requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.1.1
WTForms==3.1.2
Werkzeug==2.3.7
pytz==2024.1
email_validator==2.1.1
